specific BusInfo type for 'root hubs'
root hubs are a Linuxism - an abstract USB device capturing
information about a system USB bus; connected to a Host Controller.  Bus
data is useful for system profiling so a BusInfo type is added to crate
a more portable 'root hub' type.

Considering whether to use a HostControllerInfo highlighted the place
for root hubs. There can be multiple system USB buses on a Host
  Controller Interface (HCI), so it makes sense to have a separate type.
  Multiple buses can share the same PciInfo for example because they sit
  on the same HCI.